refactor: extract prompt refiner to separate widget
- Move `_RefinerDialog` logic to new `AIPromptRefiner` widget in `lib/widgets/refiner_panel.dart`
- Update `ControlPanelWidget` to use the new component
- Remove unused imports and legacy code in `control_panel.dart`
