+++ lib/screens/workbench/control_panel.dart
@@ -1,14 +1,11 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../../l10n/app_localizations.dart';
 import '../../services/database_service.dart';
-import '../../services/llm/llm_models.dart';
-import '../../services/llm/llm_service.dart';
 import '../../services/task_queue_service.dart';
 import '../../state/app_state.dart';
+import '../../widgets/refiner_panel.dart';
 
 class ControlPanelWidget extends StatefulWidget {
   const ControlPanelWidget({super.key});
@@ -447,8 +444,7 @@ class _ControlPanelWidgetState extends State<ControlPanelWidget> {
   void _showRefinerDialog(AppState appState, AppLocalizations l10n) {
     showDialog(
       context: context,
-      useRootNavigator: true,
-      builder: (context) => _RefinerDialog(
+      builder: (context) => AIPromptRefiner(
         initialPrompt: _promptController.text,
         selectedImages: appState.selectedImages,
         onApply: (refined) {
@@ -666,236 +662,6 @@ class _ControlPanelWidgetState extends State<ControlPanelWidget> {
           );
         },
       ),
-    );
-  }
-}
-
-class _RefinerDialog extends StatefulWidget {
-  final String initialPrompt;
-  final List<File> selectedImages;
-  final Function(String) onApply;
-
-  const _RefinerDialog({
-    required this.initialPrompt,
-    required this.selectedImages,
-    required this.onApply,
-  });
-
-  @override
-  State<_RefinerDialog> createState() => _RefinerDialogState();
-}
-
-class _RefinerDialogState extends State<_RefinerDialog> {
-  final DatabaseService _db = DatabaseService();
-  late TextEditingController _currentPromptCtrl;
-  final TextEditingController _refinedPromptCtrl = TextEditingController();
-  
-  List<Map<String, dynamic>> _models = [];
-  List<Map<String, dynamic>> _channels = [];
-  List<Map<String, dynamic>> _sysPrompts = [];
-  
-  int? _selectedModelPk;
-  String? _selectedSysPrompt;
-  bool _isRefining = false;
-  bool _isLoadingData = true;
-
-  @override
-  void initState() {
-    super.initState();
-    _currentPromptCtrl = TextEditingController(text: widget.initialPrompt);
-    _loadData();
-  }
-
-  Future<void> _loadData() async {
-    final allModels = await _db.getModels();
-    final channels = await _db.getChannels();
-    final refinerModels = allModels.where((m) => m['tag'] == 'chat' || m['tag'] == 'multimodal').toList();
-    final refinerPrompts = (await _db.getPrompts()).where((p) => p['tag'] == 'Refiner').toList();
-
-    if (mounted) {
-      setState(() {
-        _models = refinerModels;
-        _channels = channels;
-        _sysPrompts = refinerPrompts;
-        if (_models.isNotEmpty) _selectedModelPk = _models.first['id'] as int;
-        if (_sysPrompts.isNotEmpty) _selectedSysPrompt = _sysPrompts.first['content'];
-        _isLoadingData = false;
-      });
-    }
-  }
-
-  Future<void> _refine() async {
-    if (_selectedModelPk == null) return;
-    final l10n = AppLocalizations.of(context)!;
-
-    setState(() {
-      _isRefining = true;
-      _refinedPromptCtrl.clear();
-    });
-
-    try {
-      final attachments = widget.selectedImages.map((f) => 
-        LLMAttachment.fromFile(f, 'image/jpeg')
-      ).toList();
-
-      // Find the model ID string if needed, or just pass PK
-      // LLMService now accepts dynamic modelIdentifier (String or int)
-      
-      final stream = LLMService().requestStream(
-        modelIdentifier: _selectedModelPk!,
-        messages: [
-          if (_selectedSysPrompt != null)
-            LLMMessage(role: LLMRole.system, content: _selectedSysPrompt!),
-          LLMMessage(
-            role: LLMRole.user, 
-            content: _currentPromptCtrl.text,
-            attachments: attachments,
-          ),
-        ],
-      );
-
-      String accumulatedText = "";
-
-      await for (final chunk in stream) {
-        if (chunk.textPart != null) {
-          accumulatedText += chunk.textPart!;
-          setState(() {
-            _refinedPromptCtrl.text = accumulatedText;
-          });
-        }
-      }
-
-    } catch (e) {
-      if (mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(l10n.refineFailed(e.toString()))));
-      }
-    } finally {
-      if (mounted) setState(() => _isRefining = false);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    return AlertDialog(
-      title: Text(l10n.aiPromptRefiner),
-      content: ConstrainedBox(
-        constraints: const BoxConstraints(maxWidth: 800),
-        child: _isLoadingData 
-          ? const SizedBox(height: 200, child: Center(child: CircularProgressIndicator()))
-          : Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Row(
-              children: [
-                Expanded(
-                  child: DropdownButtonFormField<int>(
-                    initialValue: _selectedModelPk,
-                    decoration: InputDecoration(labelText: l10n.refinerModel, border: const OutlineInputBorder()),
-                    items: _models.map((m) {
-                      final channel = _channels.firstWhere((c) => c['id'] == m['channel_id'], orElse: () => {});
-                      return DropdownMenuItem(
-                        value: m['id'] as int,
-                        child: Row(
-                          children: [
-                            if (channel.isNotEmpty && channel['tag'] != null)
-                              Container(
-                                margin: const EdgeInsets.only(right: 8),
-                                padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 1),
-                                decoration: BoxDecoration(
-                                  color: Color(channel['tag_color'] ?? 0xFF607D8B).withValues(alpha: 0.2),
-                                  borderRadius: BorderRadius.circular(4),
-                                ),
-                                child: Text(
-                                  channel['tag'],
-                                  style: TextStyle(
-                                    fontSize: 9, 
-                                    color: Color(channel['tag_color'] ?? 0xFF607D8B),
-                                    fontWeight: FontWeight.bold,
-                                  ),
-                                ),
-                              ),
-                            Expanded(child: Text(m['model_name'], overflow: TextOverflow.ellipsis)),
-                          ],
-                        ),
-                      );
-                    }).toList(),
-                    onChanged: (v) => setState(() => _selectedModelPk = v),
-                  ),
-                ),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: DropdownButtonFormField<String>(
-                    initialValue: _selectedSysPrompt,
-                    decoration: InputDecoration(labelText: l10n.systemPrompt, border: const OutlineInputBorder()),
-                    items: _sysPrompts.map((p) => DropdownMenuItem(value: p['content'] as String, child: Text(p['title']))).toList(),
-                    onChanged: (v) => setState(() => _selectedSysPrompt = v),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 24),
-            Row(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(l10n.currentPrompt, style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 12)),
-                      const SizedBox(height: 8),
-                      TextField(
-                        controller: _currentPromptCtrl,
-                        maxLines: 10,
-                        decoration: const InputDecoration(border: OutlineInputBorder()),
-                      ),
-                    ],
-                  ),
-                ),
-                const SizedBox(width: 16),
-                const Center(child: Icon(Icons.arrow_forward, color: Colors.grey)),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(l10n.refinedPrompt, style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 12)),
-                      const SizedBox(height: 8),
-                      TextField(
-                        controller: _refinedPromptCtrl,
-                        maxLines: 10,
-                        readOnly: true,
-                        decoration: InputDecoration(
-                          border: const OutlineInputBorder(),
-                          fillColor: Colors.grey.withValues(alpha: 0.05),
-                          filled: true,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(onPressed: () => Navigator.pop(context), child: Text(l10n.cancel)),
-        FilledButton.icon(
-          onPressed: _isRefining ? null : _refine,
-          icon: _isRefining 
-            ? const SizedBox(width: 16, height: 16, child: CircularProgressIndicator(strokeWidth: 2, color: Colors.white))
-            : const Icon(Icons.auto_fix_high),
-          label: Text(l10n.refine),
-        ),
-        FilledButton(
-          onPressed: _refinedPromptCtrl.text.isEmpty ? null : () {
-            widget.onApply(_refinedPromptCtrl.text);
-            Navigator.pop(context);
-          },
-          child: Text(l10n.apply),
-        ),
-      ],
     );
   }
 }
